automations/requirements.txt
pydantic>=2.9.2,<3
openai>=1.51,<2
requests>=2.32.3,<3
pyyaml>=6.0.2,<7
beautifulsoup4>=4.12.0,<5
chromedriver-py>=129.0.6668.100
selenium>4.25.0,<5
python-frontmatter>=1.1.0,<2.0.0
translate>=3.6.1,<4.0.0